requirements.txt
langchain
flask
flask_cors
faiss-cpu
openai
PyPDF2
python-dotenv
pandas